+++ SK.CRM.Application/DTOs/Quote/QuoteItemDto.cs
@@ -10,21 +10,21 @@
 
         public string ProductName { get; set; }
         public int Quantity { get; set; } = 1;
-        public double? UnitPrice { get; set; }
+        public decimal? UnitPrice { get; set; }
 
         // Discount for this item (in percentage, e.g., 0.05 for 5%)
-        public double DiscountRate { get; set; } = 0;
+        public decimal DiscountRate { get; set; } = 0;
 
-        public double TaxRate { get; set; } = 0;
+        public decimal TaxRate { get; set; } = 0;
 
-        public double? TotalBeforeDiscount => Quantity * UnitPrice;
+        public decimal? TotalBeforeDiscount => Quantity * UnitPrice;
 
-        public double? DiscountAmount => TotalBeforeDiscount * DiscountRate;
+        public decimal? DiscountAmount => TotalBeforeDiscount * DiscountRate;
 
-        public double? TotalAfterDiscount => TotalBeforeDiscount - DiscountAmount;
+        public decimal? TotalAfterDiscount => TotalBeforeDiscount - DiscountAmount;
 
-        public double? TaxAmount => TotalAfterDiscount * TaxRate;
+        public decimal? TaxAmount => TotalAfterDiscount * TaxRate;
 
-        public double? TotalAfterDiscountAndTax => TotalAfterDiscount + TaxAmount;
+        public decimal? TotalAfterDiscountAndTax => TotalAfterDiscount + TaxAmount;
     }
 }
