+++ SK.CRM.Application/Features/Items/Dtos/ItemDto.cs
@@ -10,9 +10,9 @@
 
         public string? Description { get; set; }
 
-        public double DiscountRate { get; set; } = 0;
+        public decimal DiscountRate { get; set; } = 0;
 
-        public double TaxRate { get; set; } = 0;
+        public decimal TaxRate { get; set; } = 0;
 
     }
 }
